Change the way the websocket handles authorization
Before, it was expected that the client send an auth token as soon as the connection was opened. But sometimes the user is already in session and doesn't need to sent an auth token.

Now, the server sends a message specifically requesting an auth token when the conection is opened, if needed.

Also, messages from the server are now in a different format than messages from other clients.